WIP – refactoring config & plugins
