+++ lib/config.js
@@ -1,45 +1,29 @@
 module.exports = getConfig
+var internals = module.exports.internals = {
+  getDefaults: require('./config/defaults')
+}
 
-var fs = require('fs')
 var path = require('path')
 var url = require('url')
 
+var defaultsDeep = require('lodash').defaultsDeep
 var log = require('npmlog')
 var mkdirp = require('mkdirp')
 
 function getConfig (options) {
-  var projectPath = process.cwd()
-
-  var pkg = require(path.join(projectPath, 'package.json'))
-
+  var defaults = internals.getDefaults()
   var config = {
-    name: pkg.name,
     paths: {
-      data: options.data || path.join(projectPath, '.hoodie'),
-      public: options.public || path.join(projectPath, 'public')
+      data: options.data,
+      public: options.public
     },
-    db: {},
-    admin: {},
-    account: options.account || {}
+    app: {
+      hostname: options.bindAddress,
+      port: options.port
+    }
   }
 
-  var publicFolderExists
-  try {
-    publicFolderExists = fs.statSync(config.paths.public).isDirectory()
-  } catch (err) {
-    publicFolderExists = false
-  }
-
-  if (!publicFolderExists) {
-    config.paths.public = path.resolve(__dirname, '../public')
-    log.info('config', 'The "public" app path does not exist. Serving ' + config.paths.public)
-  }
-
-  config.app = {
-    hostname: options.bindAddress || '127.0.0.1',
-    port: options.port || 8080,
-    protocol: 'http'
-  }
+  defaultsDeep(config, defaults)
 
   if (options.dbUrl) {
     config.db.url = options.dbUrl
