+++ lib/hapi.js
@@ -33,8 +33,8 @@ function hapi (config, couchConfig, callback) {
     ]
 
     var localPlugins = [
-      require('./http-log'),
-      require('./public')
+      require('./plugins/log'),
+      require('./plugins/public')
     ].map(function (register) { return {options, register} })
 
     var hoodieCorePlugins = [{
