+++ lib/index.js
@@ -28,6 +28,18 @@ function getHoodieServer (options, callback) {
 
   mkdirp.sync(config.paths.data)
 
+  var publicFolderExists
+  try {
+    publicFolderExists = fs.statSync(config.paths.public).isDirectory()
+  } catch (err) {
+    publicFolderExists = false
+  }
+
+  if (!publicFolderExists) {
+    config.paths.public = path.resolve(__dirname, '../public')
+    log.info('config', 'The "public" app path does not exist. Serving ' + config.paths.public)
+  }
+
   var server = new hapi.Server(hapiConfig)
 
   require('./couchdb')(config, function (error, couchConfig) {
