+++ lib/plugins/log.js
