+++ lib/plugins/public.js
