+++ package.json
@@ -40,14 +40,15 @@
     "vision": "^4.1.0"
   },
   "devDependencies": {
+    "@hoodie/client": "^4.0.1",
     "bundle-bump-bot": "^1.1.0",
     "coveralls": "^2.11.9",
-    "@hoodie/client": "^4.0.1",
     "nock": "^8.0.0",
     "nyc": "^6.4.0",
     "proxyquire": "^1.7.4",
     "rimraf": "^2.5.2",
     "semantic-release": "^6.2.1",
+    "simple-mock": "^0.7.0",
     "standard": "^6.0.8",
     "tap": "^5.7.1"
   },
